#13127 refactor and cleanup
