+++ plugins/org.jkiss.dbeaver.core/src/org/jkiss/dbeaver/ui/dialogs/connection/ConnectionNameResolver.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jkiss.dbeaver;
+package org.jkiss.dbeaver.ui.dialogs.connection;
 
 import org.jkiss.code.NotNull;
 import org.jkiss.code.Nullable;
